refactor: remove unused
